+++ lib/common/routers/route_provider.dart
@@ -124,7 +124,7 @@ class MyPageRoute extends GoRouteData {
 
   @override
   Widget build(BuildContext context, GoRouterState state) {
-    return const MyPageScreen();
+    return const MyPageView();
   }
 }
