--- lib/presentation/effects/dashboard/dashboard.dart
@@ -1,1 +1,0 @@
-export 'swipe_dashboard_widget.dart';
