+++ lib/presentation/effects/effects.dart
@@ -1,3 +1,2 @@
 export 'camera_quickshot/camera_quickshot.dart';
-export 'dashboard/dashboard.dart';
 export 'emojis/emojis.dart';
