+++ lib/presentation/my_page/screens/screens.dart
@@ -1,2 +1,2 @@
 export 'add_resolution_screen.dart';
-export 'my_page_screen.dart';
+export 'my_page_view.dart';
