--- lib/presentation/my_page/widgets/resolution_accomplishment_donut_graph_datamodel.dart
@@ -1,9 +1,0 @@
-class ResolutionAccomplishmentDonutGraphDataModel {
-  ResolutionAccomplishmentDonutGraphDataModel({
-    required this.x,
-    required this.y,
-  });
-
-  final dynamic x;
-  final num? y;
-}
