+++ lib/presentation/presentation.dart
@@ -14,8 +14,6 @@ export 'effects/camera_quickshot/camera_quickshot.dart';
 export 'effects/camera_quickshot/reaction_camera_widget.dart';
 export 'effects/camera_quickshot/reaction_camera_widget_model.dart';
 export 'effects/camera_quickshot/reaction_camera_widget_model_provider.dart';
-export 'effects/dashboard/dashboard.dart';
-export 'effects/dashboard/swipe_dashboard_widget.dart';
 export 'effects/effects.dart';
 export 'effects/emojis/animation_sample_page.dart';
 export 'effects/emojis/balloon_animation/balloon_animation.dart';
@@ -70,13 +68,8 @@ export 'my_page/providers/add_resolution_provider.dart';
 export 'my_page/providers/my_page_resolution_list_provider.dart';
 export 'my_page/providers/providers.dart';
 export 'my_page/screens/add_resolution_screen.dart';
-export 'my_page/screens/my_page_screen.dart';
+export 'my_page/screens/my_page_view.dart';
 export 'my_page/screens/screens.dart';
-export 'my_page/widgets/resolution_accomplishment_donut_graph_datamodel.dart';
-export 'my_page/widgets/resolution_dashboard_widget.dart';
-export 'my_page/widgets/resolution_doughnut_graph_widget.dart';
-export 'my_page/widgets/resolution_linear_gauge_graph_widget.dart';
-export 'my_page/widgets/widgets.dart';
 export 'reaction/provider/provider.dart';
 export 'reaction/provider/reaction_animation_widget_state_provider.dart';
 export 'reaction/reaction.dart';
